[lab10]  исправил замечания

+++ lab10/src/main/kotlin/ViewModel/Point.kt
@@ -1,3 +1,3 @@
 package ViewModel
 
-data class Point(var x: Int, var y: Int)
+data class Point(var x: Float, var y: Float)
